requirements.txt
amqp==5.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.2.3
Django==4.2
django-js-asset==2.0.0
django-mptt==0.14.0
django-stubs==1.16.0
django-stubs-ext==0.8.0
djangorestframework==3.14.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
iniconfig==2.0.0
kombu==5.2.4
mypy==1.1.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
pytest==7.3.1
pytest-celery==0.0.0
pytest-django==4.5.2
pytest-mock==3.10.0
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.4
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.9
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
